Modifications to allow building of toolset in OpenBSD

+++ src/stlink-lib/flash_loader.c
@@ -788,7 +788,7 @@ int32_t stlink_flashloader_write(stlink_t *sl, flash_loader_t *fl, stm32_addr_t 
       if (stm32l1_write_half_pages(sl, fl, addr, base, len, pagesize)) {
         return (-1);
       } else {
-        off = (size_t)(len / pagesize) * pagesize;
+        off = (uint32_t)(len / pagesize) * pagesize;
       }
     }
